WIP splitting into two table views
